refactor(onboard): isolate N1x rebuild eligibility
