+++ src/lib/actions/sandbox/rebuild-target-preflight.ts
@@ -19,4 +19,5 @@ export {
   hydrateMessagingConfigForRebuild,
   prepareRebuildRecreateOptions,
   stageRebuildHermesDashboardConfig,
+  stageRecordedManagedVllmIntent,
 } from "./rebuild-target-staging";
